Add hptool info option to show versions & packages that will be included
Since the hptool contains within its executable, the HP release number
and the exact versions of packages to be included in the resulting HP
release it will construct, a way to query a given hptool is needed.
This commit adds a "--info" (short form "-i") as an option to the hptool
command line.  The info included is the HP version to be built, and the
name and version of each package.  The exact way the info is printed is
reasonable but arbitrary with this commit, and could easily be changed.