Fix localCommand for Windows platform
The 'localCommand' function looks for an executable by looking for a file
of the exact command name, which usually will fail on Windows platforms
since the filenames for executables end with ".exe".  This change uses
Development.Shake.FilePath support to include the ".exe" when appropriate.

This change should be transparent for non-Windows platforms, and is
necessary, but not sufficient, for the HP build on Windows.