Put uninstaller info in correct registry in 64-bit
This broken with the changes to detect previous installs.
The problem is minor and only affects users on 64-bit
Windows, who try to install when there are other HP's
installed, and they select the "install anyway; I know what
I'm doing" path.  And it was still uninstallable, even in the
wrong "side" of the registry (the 32-bit side rather than
64-bit side).

This file is used to generate the installe executable so it
is the last step in the build and thus the effect on anything
else should be nothing.