Switch to Shake 0.14 or later
Shake renamed *> to %> because of the upcoming *> in Prelude.
Renamed our */> to %/> to match.

The type of CmdResults was generalized in 0.15, and the code
needed to fix the type in a place where we ignored Stderr output.
The fix is just an extra type annotation, so should work with
older Shake as well.