Rework of Haskell Platform website
This is a rework of the Haskell Platform website.

This work began as an adaptation of Erin Depew's front-end design. It
was the consensus that this design, while attractive and very usable,
could try harder to quickly bring the potential user to a download. In
particular, we should meet the standard set by Scala [1] (and others),
where the user can get from the http://scala-lang.org/ to a downloaded
installed in two clicks.

Towards this goal, the Haskell Platform index page has been eliminated
entirely.  This is possible as haskell.org now does an admirable job of
selling the language, thanks to Chris Done's rework. The download page
now initially shows only those downloads for the user's current
operating system, while allowing the user to explore other options if
neccesary.