Un-overload the "targetDir" target
If a FilePath matches a so-called "phony" target, Shake will
count this as satisfying the phony target and skip its steps.
Currently, the target called targetDir is overloaded, being
used in a "~>" rule (for defining a "phony" target) but it is
also used for contructing file paths, and it appears as a
dependency in some rules (e.g., "need [targetDir]"). When it
is a dependency, it actually should be upon a phony target,
not the actual directory.  Due to how dependencies are
constructed for the Windows HP build, this problem has
plagued it.

This change needs to be tested for the other build environments.

* hptool/src/Main.hs
* hptool/src/OS/Internal.hs
* hptool/src/OS/Mac.hs
* hptool/src/OS/Posix.hs
* hptool/src/OS/Win.hs
* hptool/src/OS/Win/WinNsis.hs
  * change the dependency upon targetDir to phonyTargetDir
* hptool/src/Paths.hs
  * define and export phonyTargetDir
* hptool/src/Target.hs
  * Change rule from targetDir to phonyTargetDir