Work around 2GB payload limit of NSIS
The installer framework currently used for HP (the open
source project NSIS) uses a 32-bit-centric file format
(apparently) and thus has a limitation of 2GB as the maximum
size of a blob that can be included.  After investigating
numerous alternatives, the simplest work-around seemed to be
just to bundle and silently install a number of child
installers, breaking the actual HP payload into three
logical components: GHC, MSys, Stack, and the HP-specific
pre-built packages.

Three additional installer scripts are added here, to build 3
of the new child-installers (stack is already packaged with
its own NSIS installer, so that is happily used unmodified),
along with the necessarily bookkeeping to populate an NSI
file from a template, and shake rules to build these
installers at the right time with needed contents.  Further,
the original NSI file is modified to add some new UI (for
stack) and create the NSI install-time actions to launch the
new child-installers as necessary.


* hptool/os-extras/win/templates/Extralibs.nsi.mu
  * New. Template (using the "Mustache" syntax) for the
    additional pre-built HP-installed packages.
* hptool/os-extras/win/templates/GHC.nsi.mu
  * New. Template (using the "Mustache" syntax) for
    everything that is part of the GHC pre-built binary
    distribution.
* hptool/os-extras/win/templates/MSys.nsi.mu
  * New. Template (using the "Mustache" syntax) for the
    MSys content.
* hptool/os-extras/win/templates/Nsisfile.nsi.mu
  * In addition to the changes already mentioned above, some
    white space clean up to make it easy to read.  Add a
    public-service annoucement at the end of the installation
    to explain to users that they may see a brief "pop-up"
    window immediately after closing the installer (it is ghc
    running to update the package spec; this action is not
    new but the PSA is).
* hptool/src/OS/Win.hs
  * Build rules for the new sub-installers
* hptool/src/OS/Win/WinNsis.hs
  * The rules for building the NSI file from the template
* hptool/src/OS/Win/WinPaths.hs
  * Paths for all the new sub-installer templates, output
    files, etc.
* hptool/src/OS/Win/WinRules.hs
  * Replace the rule invocations for the NSIS
    generation. For stack, the linux-ish platforms simply
    copy the stack executable but for Windows, there is an
    installer and it is launched as part of the HP install,
    so remove the action here to copy the stack executable.