Improve silent install functionality
Add a command line option to the HP installer executable so
that an installation path for stack can be specified.
Previously, an installation path could be specified for the
Haskell Platform (ghc, et al), but not for stack (and stack's
location cannot be inferred from the HP default location).
With this change, both the HP components and stack can be
silently installed to separate, user-specified locations.

* hptool/os-extras/win/templates/Nsisfile.nsi.mu
  * Add support for a command-line option to the installer to
    allow a parameter, "/STACK=", to specify the installation
    path for stack
  * Improve the silent-install behavior for 64-bit versus
    32-bit installations
* hptool/src/OS/Win/WinNsis.hs
  * Update the size of the installed stack package