Workaround cabal #5362: "cabal clean" behavior 2.2 vs 2.0
Using cabal-install 2.0, if a "cabal clean" is done on a
non-existent directory, it will successfully complete.  In
cabal-install 2.2, "cabal clean" on a non-existent directory
will fail (on Windows, at least) with "dist\build:
getDirectoryContents:findFirstFile: does not exist (The
system cannot find the path specified.)"  So, we will
temporarily work around this by creating the directory if it
doesn't already exist, before we invoke "cabal clean".

* hptool/src/Target.hs
  * use createDirectoryIfMissing (via Util.makeDirectory) to
    create the directory about to be "cleaned" if not already
    there, to work-around the cabal issue.